fs: Add a header file for I/O iteration
Provide a struct iov_iter and some related functions to encapsulate an
iterator used when reading and writing files. This will be used with the
new 'file' uclass.

While it is modelled on Linux, I/O vectors have been omitted for now.
These can be added later without updating the interface used by the
uclass.

Signed-off-by: Simon Glass <sjg@chromium.org>