sandbox: Enable the new virtio emulator
Add the devicetree snippet to include the new virtio block-device
emulator in the test devicetree.

Series-to: u-boot
Cover-letter:
virtio: Add a fully functional virtio emulator
The current implementation of virtio in sandbox is fairly basic. It is
enough to test the behaviour of queues, but it does not test the full
stack, e.g. using MMIO to access a block device.

This series adds a new type of virtio emulator which is capable of
handling MMIO. Some support for MMIO is added to sandbox in the process.
With this, a dummy block device can be used from within U-Boot, without
running on QEMU
END

Signed-off-by: Simon Glass <sjg@chromium.org>