scsi: Use REPORT LUNS to scan the correct number of LUNs
The current algortihm scans all LUNs on all targets. Where the potential
number is very large, as with QEMU (256 targets each with 4K LUNs) this
is a substantial waste of time.

Use the REPORT LUNS command to determine the number of LUNs on each
target. Where that fails, fall back to scanning everything.

Signed-off-by: Simon Glass <sjg@chromium.org>