scsi: Tidy up error returns for read/write
The current code never returns an error if something goes wrong. While
it is normally useful to return the number of blocks processed, when
nothing is read/written at all, there is clearly something wrong.

Update the logic to return -EIO in this case.

Avoid changing blkcnt since it is an input parameter and this is
confusing.

Signed-off-by: Simon Glass <sjg@chromium.org>