emulation: Introduction MACH_QEMU
We have QEMU targets across various architecture. So there is not really
a QEMU archicture. So 'ARCH_QEMU' seems a bit peculiar.

We have a board/emulation directory where much of the arch-specific code
is kept. But QEMU is not really a board; in fact many boards use QEMU.

While 'emulation' is a useful concept, QEMU is only one of the emulators
available. so MACH_EMULATION seems too vague to capture QEMU-specific
features.

On ARM the code is in the arch/arm/mach-qemu directory and it seems
reasonable to go with this approach and consider QEMU to be a machine.

So add a MACH_QEMU option. For now nothing uses it.

Signed-off-by: Simon Glass <sjg@chromium.org>