acpi: Move code table-writing code to lib/acpi/acpi.c
The acpi_table.c file includes ACPI_WRITER() declarations and is only
included in the build if CONFIG_ACPIGEN is enabled.

For QEMU we want to augment the tables provided via qfw, but
CONFIG_ACPIGEN is not enabled.

Move the core functions out to the acpi.c file which is included in all
x86 builds. This allows the QEMU build to make use of these functions.

Signed-off-by: Simon Glass <sjg@chromium.org>