boot: Create a function for final pre-boot steps
There are various functions which announce that booting is imminent and
do related preparation. Most of these are arch-specific.

In practice, most archs do the a similar thing. It would be better to
have a common function, with perhaps some events for things that are
really arch- and board-specific.

Create a new function for this. For now, nothing uses it.

Signed-off-by: Simon Glass <sjg@chromium.org>