virtio: Correct dir creation to use existing stream
Rather than creating a new stream, which gets lost, virtio_fs_dir_open()
should use the existing one. Fix it, so that the fh member is preserved
for future use.

This fixes a problem where the fh value is not preserved, thus causing
a subsequent 'ls' on the same directory to fail.

Signed-off-by: Simon Glass <sjg@chromium.org>
Fixes: #15