test/py: hooks: Provide a proper way to restart U-Boot
Board reset in pytests is pretty ad-hoc at present. The normal approach
is to cause a reset to happen, or come to a point in a test where a
reset is expected. Then the test looks for the U-Boot banner, indicating
that the board has rebooted.

This has a few problems:
- Some lab boards use USB to send U-Boot to the board. Resetting may
  result in a hang, since U-Boot is not present after a reset
- Resetting generally does a software reset, which is different to a
  hardware reset. For example the TPM is not reset on a software reset
  so the previous TPM state is still present. This means that tests
  which want to check early TPM-setup cannot work
- To do a hardware reset, the current integration simply throws away the
  ConsoleExecAttach() object and created a new one. This unnecessary
  with the sjg Labgrid integration and it is quite inefficient, since
  U-Boot is built again, written to the board again, etc.
- Some boards may need a signal to be provided to indicate whence the
  firmware should be read. If this signal happens to not be present, the
  wrong firmware will be started

There are various other corner cases as well.

To resolve this, provide a restart_board() implementation in
ConsoleExecAttach which can reliably restart a board. There is one
parameter, which tells it whether the reset has already been triggered.
If not, it triggers one.

Implement this in the sjg Labgrid integration (i.e. 'lab mode', via a
new `u-boot-test-restart` hook. Enable it for ellesmere (sjg lab).

This allows send-only boards like pcduino3 to handle the EFI watchdog
test, for example.

There is quite a bit of cleanup which could be pursured following this
change, mostly calling the new function instead ubman.run_command() and
improving how sandbox restarts. This can be dealt with later.

Series-to: concept
Cover-letter:
test/py: Various enhancements for lab use
This series provides several improvements associated with the sjg labgrid
integration:

- restart boards reliably during tests
- console logging to a file
- allow putting Labgrid into verbose mode
- easier debugging of test.py
END

Signed-off-by: Simon Glass <sjg@chromium.org>