test: Convert FIT test into separate pieces
Use a separate function for each test, so that they appear separately
in the results.

Move the comment about what tests are covered to the top.

Create an 'fsetup' fixture which provides what each test needs.

Adjust the helpers to fit better with this new scheme.

This should make the test more straightforward and easier to extend in
future.

Signed-off-by: Simon Glass <sjg@chromium.org>