test: Add a test for loading a second FIT
Check that we can read a FIT containing a devicetree, then boot with a
second FIT.

This requires a few more FIT parameters and commands to be configurable.

Series-to: concept
Series-cc: heinrich
Cover-letter:
Support FITs with load-only configs
FIT is growing a new feature where a load-only configuration can be used
to obtain and set up a devicetree prior to loading the OS.

This series implements this feature in U-Boot.

Some refactoring is necessary to permit this, on top of a previous
series. Also the FIT tests have the annoying property that they all run
at once, so this series also adjusts these.

Further work will integrate this properly into a bootmeth, etc.
END

Signed-off-by: Simon Glass <sjg@chromium.org>