boot: Support bootm restart
Normally bootm proceeds sequentially through the various states, from
'start' to 'go'.

In some cases we want to load the devicetree from one FIT and the kernel
and ramdisk from another. This requires two bootm commands.

Of course it is possible to just do a second 'bootm start' to load the
kernel. But that removes all record of the devicetree from the
boot_images information, so it is then not provided to the OS.

Add a 'restart' subcommand which allows more images to be loaded,
without erasing those already loaded.