boot: Update boot_get_ramdisk() to return an error
This function returns 1 on error (at least in some cases). Change it to
use an error code, so that more information can be provided. For now it
mostly returns either 0 or -ENOPKG

Rename the return arguments, dropping the rd_ prefix and adding a 'p'
(for pointer) suffix to return arguments so that is clear that this is
what they are.

Update the comments for clarity and to fit with the usual style. Avoid
documenting the return arguments in two places.

Provide a common path for all error returns.

Signed-off-by: Simon Glass <sjg@chromium.org>