boot: Update boot_get_ramdisk() return arguments
If an error is returned, leave the return arguments alone. There is no
point in setting them to zero, since the caller already knows (from the
error code) that they are not being returned.

This makes things simpler for callers which have an existing ramdisk.

Handle the -ENOPKG error code in bootm_find_images()

Signed-off-by: Simon Glass <sjg@chromium.org>