test: Convert fs_helper to use a class
It is somewhat inconvenient that callers to mk_fs() must create their
own temporary directory. Convert it to a class so this can be handled
automatically, using a context manager.

Rather than specifying the file size in bytes, use MB since it is rare
to need a smaller file.

Update mk_fs() to use this new class.

Signed-off-by: Simon Glass <sjg@chromium.org>