boot: pxe: Add a token for FIT
It is not obvious that PXE supports FIT, but it does, by detecting
whether an image is a FIT or not.

In many cases a FIT is more convenient than using separate files for
the kernel, initrd and devicetree.

Really we should promote FIT as an important format, rather than
silentily dealing with it if detected.

Add a new 'fit' token which indicates that a FIT is being used.

When the 'fit' token is used, the expectation is that the devicetree
is within the FIT, but this is not required, for now.

Signed-off-by: Simon Glass <sjg@chromium.org>