CI: Checking building of extended boards
Enable building of extended boards in the world build, so we can ensure
these don't collect failures over time.

For now, manually add a buildman rule for riscv to avoid this error:

   qemu-riscv64_smode   Tool chain for 'riscv' is missing: do nothing

Series-to: concept
Series-cc: heinrich
Cover-letter:
buildman: Introduce some order to config fragments
U-Boot supports config fragments, separate files which can be used to
amend the config for an existing board.

So far, buildman has not supported this feature. It is not been possible
to buildman boards with fragments, nor has there been any file which
describes which fragments are valid with which boards.

This series attempts to improve this situation. It updates buildman in
two ways:

- Provides an option to pass fragment files to the buildman build
- Provides a file which associates boards with supported fragments

With this, extended boards are built in CI (Gitlab only).
END

Signed-off-by: Simon Glass <sjg@chromium.org>