video: efi: Store the GOP in priv data
At present the EFI GOP is read in the probe() method but is not stored,
so it is not possible to use it later.

With the EFI app we need to be able to access it after probing, when
blitting is used. Create a struct to hold it. Store the framebuffer
pointer in there as well, to simplify the code.

Signed-off-by: Simon Glass <sjg@chromium.org>