boot: Enable the VBE-OS bootmeth for EFI app and payload
This method is useful for locating an OS using the VBE approach. Enable
it for the app and payload, for both x86 and ARM.

Series-to: concept
Series-cc: heinrich
Cover-letter:
efi: Support video output on ARM
So far the EFI app supports video on x86 but not ARM. This series adds
this feature, using EFI GOP's 'blt' feature.

Support for bootstd is enabled at the same time, so that the app can
boot an OS automatically.
END

Signed-off-by: Simon Glass <sjg@chromium.org>