boot: Update BOOTM_EFI to depend on EFI_LOADER
The EFI loader uses a rather unpleasant global for starting images. For
now we only intend to support the EFI bootmeth from the app, so update
the condition so that booting an EFI app is only possible from the
EFI loader.

Future work may enable the 'bootefi' command for the app.

Signed-off-by: Simon Glass <sjg@chromium.org>