boot: Update BOOTM_EFI to depend on EFI_LOADER
The EFI loader uses a rather unpleasant global for starting images. For
now the only way we intended to support booting EFI apps from the U-Boot
EFI app is via a bootmeth.

Update the condition so that the 'bootefi' command only works from the
EFI loader.

Future work may enable the 'bootefi' command for the app.

Signed-off-by: Simon Glass <sjg@chromium.org>