boot: Allocate the kernel buffer when running in the app
The app cannot know the address to use in advance, so does not have a
value for kernel_addr_r in its environment, if it even has an
environment.

Allocate memory for the kernel instead. Limit it to 64MB which seems
like plenty for the distros out there. It is enough to load a full
kernel if needed.

Signed-off-by: Simon Glass <sjg@chromium.org>