boot: Support reserving memory in get_relfile_envaddr()
We normally require environment variables when loading files using PXE/
syslinux. This is not really necessary, since we can use lmb to find a
suitable space.

Add support for reserving space in get_relfile_envaddr(), which is the
commonly used function for reading a file. Use env_get() so no error is
shown if the environment variable does not exist.

Signed-off-by: Simon Glass <sjg@chromium.org>