efi: Use variables in dp_size()
This function calculates the parent size in many places and does the
same addition in most cases. Create a variable for the parent_size and
another for the device tree, so we can do the addition once a the end.

The uclass array is too sparse to consider an array.

Signed-off-by: Simon Glass <sjg@chromium.org>