efi_client: Provide generic malloc() and free() functions
The device-path code requires memory allocation. Provide an
implementation of these for use by the client.

Note that this will not work for the stub, since in that case it needs
to use U-Boot's memory allocation. This was a subject of a long argument
on the mailing list about U-Boot using its own malloc() where possible,
so it seems best to just leave this broken.

Signed-off-by: Simon Glass <sjg@chromium.org>