efi: arm: Allow --gc-sections in the EFI app
This option is used on most other boards. Leaving it disabled creates
situations where the linker fails to build, due to symbols which are
referenced only it code that would have been garbage-collected.

To keep this disabled we would need to introduce #ifdefs not needed by
other platforms, which would be annoying.

The only current reason why this cannot be enabled is that the
linker lists are lost. Add a KEEP() around these, and enable the option.

Signed-off-by: Simon Glass <sjg@chromium.org>