efi: Adjust serial to work better on real devices
The EFI serial driver checks for keypresses but does not allow other
events to be processed. For devices which have a USB keyboard, the USB
message must be processed before any keys can emerge for the serial
driver to pick up.

Add a periodic timer to the list of events. This allows the keyboard to
work on Qualcomm laptops, for example.

Series-changes: 2
- Add a comment to serial_efi_get_key() about the 1ms delay

Signed-off-by: Simon Glass <sjg@chromium.org>