boot: Support lmb reservation in resolve_os_comp_buf()
This function is very simple at present, only supporting a fixed buffer
of a fixed size.

For cases where the address is not provided, we may still want to set a
limit on the size.

Support lmb reservation of an existing buffer as using lmb to reserved a
new buffer. Use the provided size information to select a suitable size
for the decompression buffer.

Signed-off-by: Simon Glass <sjg@chromium.org>