timer: Mark this as incompatible with SYS_ARCH_TIMER
The timers implemented under SYS_ARCH_TIMER are not part of the TIMER
framework and thus cannot be enabled at the same time. Express this
dependency in Kconfig and add a note to the help of SYS_ARCH_TIMER.

Reported-by: Marius Dinu <m95d+git@psihoexpert.ro>
Signed-off-by: Tom Rini <trini@konsulko.com>
---
Marius, I ended up reworking things a bit (and also found part 1/2 of
this series) but please let me know if you feel this should retain your
Author and such instead. Thanks!