efi: Drop EFI_CALL() from efidebug command
The intend of this macro is to debug calls from an app. But the efidebug
command is not an app, so it really doesn't make a lot of sense to call
via that macro. Remove EFI_CALL() from do_efi_show_handles() so that the
app can use this code.

Signed-off-by: Simon Glass <sjg@chromium.org>