efi: Update path_uefi() to avoid allowed_unaligned()
This call seems to have been added in an unrelated commit, likely to fix
a bug:

  046fe7b5074 ("efi_loader: efi_dp_from_file() expect UTF-8 path")

With standard boot we don't need to call efi_set_bootdev(). Also this
function is not available in the EFI app.

So for now, add a condition, to avoid a build error in the x86 app when
CONFIG_EFI_BINARY_EXEC is enabled.

Signed-off-by: Simon Glass <sjg@chromium.org>