test: vbe: Avoid using fdt_util to compile a devicetree
The VBE test uses fdt_util to compile a devicetree.

Other tests log the output of the dtc tool, which can be helpful when
a test fails.

For reasons which are unclear, using fdt_util means that the path to dtc
ends up being './scripts/dtc/dtc' when run from 'make qcheck'. This
causes a failure to set up the image required for the VBE test.

Update the test to use the normal run_and_log() method instead.

Signed-off-by: Simon Glass <sjg@chromium.org>