test/py: Update test_ut_dm_init_bootstd() to avoid ubman
Pass in the two fixtures this function actually requires, rather than
ubman, which is a function-scoped fixture.

Series-to: concept
Cover-letter:
test: Move towards using a fixture for test images
There are quite a few test images created by the test system. Most are
needed for bootstd tests.

At present these are created by a test, called test_ut_dm_init_bootstd()
which runs before the test_ut() tests.

This means that it is possible to create the images once, then iterate
on a test without the images being recreated each time.

While this is not a terrible approach, it does have some drawbacks:

- If an image is missing or corrupted for some reason, the test fails
- The relationship between the images and the tests is not obvious
- Images are often created even if the tests to be run do not need them

Overall, it seems worthwhile to explore moving to using a fixture for
the images.

Another problem is that 'make qcheck' and 'make pcheck' are not tested
in CI, so can break over time. The latter is a good way of seeing
whether each test is independent, e.g. not reliant on files created by
another test.

Overall this area of U-Boot could use some attention. This series makes
a very small start on this:

- Update test_fdt_add_pubkey() to work without test_vboot_base()
- Fix 'make qcheck' so that it passes
- Minor fixes to a few tests
- Pass separate fixtures to the image-setup functions instead of the
  overall ubman fixture
END

Signed-off-by: Simon Glass <sjg@chromium.org>