sandbox: Move PCI EA space out of RAM
The address chosen for testing this feature is 1MB which is part of the
sandbox RAM. When devices access this, e.g. with map_sysmem(), the
memory is mapped to a PCI device. Any changes then apply to that device
and are not written to memory. Reads also come from the device.

It is not safe to use RAM space in this way.

A symptom that something is wrong is the log message:

   map_physmem: Warning: partial map at 100000, wanted 4, got 2000

Move the memory out of the way and document it.

Signed-off-by: Simon Glass <sjg@chromium.org>
Fixes: 21ebbafde8d ("test: dm: Add a test for PCI Enhanced Allocation")